backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest-django==4.5.2
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
